Interpolation: Remove methods longer supported
The method getMultipleInterpolants no longer works in our context.
It assumed only original clauses are in leaves, but this has not been
true for a long time. It has always been true only in propositional
context, but never in the context of SMT.

All other removed methods were calling getMultipleInterpolants
eventually.

We can bring back the API if needed after more cleanup of the
interpolation code and API.