Common: write a common wrapper for assert
Currently `assert` is coming from `#include <assert.h>` or from
`#include <cassert>`.  This commit introduces a layer that either takes
`assert` from `cassert`, or depending on a new `cmake` option
`OSMT_ASSERT_THROWS`,
optionally defines `assert` as something that throws.

After this fix we no longer can use `assert` in functions specified as
`nothrow`.  This is not a problem in the current code.

Co-authored-by: Antti Hyvarinen <antti.hyvarinen@gmail.com>